// functions/src/models/userRef.ts
/**
 * Users API
 * Users API for the RecruitBuddy project
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Address } from './address';
import { Nationality } from './nationality';
import { Phone } from './phone';
import { NameSuffixEnum } from './user';

export interface UserRef { 
    userID?: string;
    userType?: string;
    firstName?: string;
    middleName?: string;
    lastName?: string;
    displayName?: string;
    nameSuffix?: NameSuffixEnum;
    preferredName?: string;
    title?: string;
    addresses?: Array<Address>;
    dateOfBirth?: string;
    nationality?: Nationality;
    phones?: Array<Phone>;
    emailAddress?: string;
    userResumesRef?: string;
    userTemplatesRef?: string;
    password?: string;
}
/*export namespace UserRef {
    export type NameSuffixEnum = 'Mr.' | 'Ms.' | 'Mrs.' | 'Miss';
    export const NameSuffixEnum = {
        Mr: 'Mr.' as NameSuffixEnum,
        Ms: 'Ms.' as NameSuffixEnum,
        Mrs: 'Mrs.' as NameSuffixEnum,
        Miss: 'Miss' as NameSuffixEnum
    };
}*/