// functions/src/models/user.ts
/**
 * Users API
 * Users API for the RecruitBuddy project
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Address } from './address';
import { Nationality } from './nationality';
import { Phone } from './phone';
import { UserResume } from './userResume';
import { UserTemplate } from './userTemplate';

export interface User { 
    userID?: string;
    userType?: string;
    firstName?: string;
    middleName?: string;
    lastName?: string;
    displayName?: string;
    nameSuffix?: NameSuffixEnum;
    preferredName?: string;
    title?: string;
    photoURL?: string,
    addresses?: Address;
    dateOfBirth?: string;
    nationality?: Nationality;
    phones?: Array<Phone>;
    email?: string;
    website?: string;
    profileText?: string;
    preferredBillingCurrency?: string;
    preferredLanguage?: string;
    userResumes?: Array<UserResume>;
    purchasedTemplates?: Array<UserTemplate>;
    userSuppliedResume?: string;
}
//export namespace User {
    export type NameSuffixEnum = 'Mr.' | 'Ms.' | 'Mrs.' | 'Miss';
    export const NameSuffixEnum = {
        Mr: 'Mr.' as NameSuffixEnum,
        Ms: 'Ms.' as NameSuffixEnum,
        Mrs: 'Mrs.' as NameSuffixEnum,
        Miss: 'Miss' as NameSuffixEnum
    };
//}